requirements.txt
streamlit==1.31.1
pillow==10.2.0
mlflow==2.10.2
torch==2.2.0
torchvision==0.17.0
lightning==2.2.0.post0